ref(lint): better, improved, cleaner linter
ChartValidation struct keeps track of a linting instance.
Validation struct keeps track of a validation and its
dependent validations as well. This will help us link validations
in a proper format and be able to more easily see and define
validations rather than rely on logic nested in loops and
arbitrary functions.